PlayerAttack Functionality and Animation
Added the Slash Sprite and made animation for it
Made the slash collidable with enemies and added the functionality for it to do damage to the enemy
Additionally fixed some name conventions and changed some types if variables to better suit the project

+++ Assets/Scripts/Classes/Enemy.cs
@@ -9,9 +9,9 @@ namespace Assets.Scripts.Classes
 {
     public class Enemy : MonoBehaviour
     {
-        public double Hp;
-        public double Str;
-        public double Def;
+        public float Hp;
+        public float Str;
+        public float Def;
         public float Speed;
         public bool Friendly;
         public bool Alive;
